@carlosonunez Workaround for Docker Compose/.env circular dependency
Docker Compose requires that all `env_files` that are declared in
`docker-compose.yml` be present on the filesystem before running any
commands like `run` or `up`.

This becomes problematic when we decrypt our environment file (where AWS
settings live) because:

- That information is stored in `env.gpg`, which gets decrypted into
`.env`, and

- That decryption will fail if `.env` is already present.

This is a circular dependency.

Separating the `gpg` Docker Compose services into their own file and
telling Compose to use that resolves this because that file does not
have references to `env_file`.